+++ cdk.context.json
@@ -1,0 +1,6 @@
+{
+  "acknowledged-issue-numbers": [
+    34892,
+    34892
+  ]
+}
